Consolidate redundant cache helper functions
